+++ src/main/kotlin/no/nav/syfo/application/ApplicationServer.kt
@@ -4,10 +4,9 @@ import io.ktor.server.engine.ApplicationEngine
 import java.util.concurrent.TimeUnit
 
 class ApplicationServer(
-    private val applicationServer: ApplicationEngine
+    private val applicationServer: ApplicationEngine,
+    private val applicationState: ApplicationState
 ) {
-    private val applicationState = ApplicationState()
-
     init {
         Runtime.getRuntime().addShutdownHook(Thread {
             this.applicationServer.stop(TimeUnit.SECONDS.toMillis(10), TimeUnit.SECONDS.toMillis(10))
